pyproject.toml
[project]
name = "software-kvm"
version = "0.1.0"
description = "Cross-platform KVM software using CH9329 and video capture card"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Mark"}
]
keywords = ["kvm", "remote-control", "video-capture", "ch9329"]

dependencies = [
    "PyQt6>=6.6.0",
    "pyserial>=3.5",
    "pynput>=1.7.6",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
software-kvm = "src.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
